move `tbot` subpackages to `lib/tbot`
