+++ lib/tbot/botfs/botfs.go
