+++ lib/tbot/botfs/botfs_test.go
