+++ lib/tbot/config/config_test.go
