+++ lib/tbot/identity/artifact.go
