+++ lib/tbot/identity_test.go
@@ -22,8 +22,8 @@ package tbot
 import (
 	"testing"
 
-	"github.com/gravitational/teleport/tool/tbot/config"
-	"github.com/gravitational/teleport/tool/tbot/identity"
+	"github.com/gravitational/teleport/lib/tbot/config"
+	"github.com/gravitational/teleport/lib/tbot/identity"
 	"github.com/gravitational/trace"
 	"github.com/stretchr/testify/require"
 )
